Fix: -Owner Summon animation. -pvp damage in pease zone. -l2off like npc interaction rotate the player to face the npc -The infamous "Air" phx script fix. -CastleManager npe. -onMagicFinalizer npe -attacking monsters circling the player removed rnd and added a check to avoid nearby mobs around the player without any sense.
-player speed to synchronize with server-client.
-when using delux key if there is no Success chests attack before decaying.
-deathpenalty and player exp percent lost close enough to official,removed delevel config, and now 
player lvl > 9 and without lucky skill, on die loosing exp. 
there are some exceptions for example: can't lose exp in pvparena-siege participants inside siege zone etc.
also added a emum(PlayerExpLost) to avoid hardcode.

Rework:
ItemsAutoDestroy
SkillTreeData
FleeingNPCs are correctly working(Rabbit-Elpy)

Clean up:
SpawnTable
L2Spawn:removed spawnlistener
HitMan config.
Champion config delete, champion are now static in core(config.java) you can modify it there.
NpcBufferTable
NpcBufferInstance
BuffTemplateTable
NpcBufferSkillIdsTable

Quests:

Added:
Config to enable-disable FakePlayer system
Q114_ResurrectionOfAnOldManager, thanks to BossForever.
Q333_HuntOfTheBlackLion, thanks to BossForever.

Q334_TheWishingPotion.
Q335_TheSongOfTheHunter.
Q336_CoinsOfMagic.

+++ L2JHellasC/java/com/l2jhellas/gameserver/network/serverpackets/GMViewCharacterInfo.java
@@ -16,7 +16,7 @@ public class GMViewCharacterInfo extends L2GameServerPacket
 	@Override
 	protected final void writeImpl()
 	{
-		float moveMultiplier = _activeChar.getMovementSpeedMultiplier();
+        double moveMultiplier = _activeChar.getMovementSpeedMultiplier();
 		int _runSpd = (int) (_activeChar.getRunSpeed() / moveMultiplier);
 		int _walkSpd = (int) (_activeChar.getWalkSpeed() / moveMultiplier);
